requirements.txt
openai>=1.0.0
pypdf>=3.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
tqdm>=4.65.0
pdfplumber>=0.9.0